layer.conf: add wrlinux specific configs
Issue: LINCD-8843

LLVMVERSION is used in oe-core. clang layer override it to ensure all
the llvm related is provided by this layer. llvm is provided by recipe
clang in this layer.

This cause warning like:
NOTE: preferred version 14.0.1 of llvm not available (for item llvm)
NOTE: versions of llvm available: 13.0.1

(LOCAL REV: NOT UPSTREAM) -- this fix is wrlinux specific

Signed-off-by: Changqing Li <changqing.li@windriver.com>